fix install so that it doesn't overwrite a backup with a symbolic link

+++ install.sh
@@ -3,7 +3,7 @@ function link_file {
     source="${PWD}/$1"
     target="${HOME}/${1/_/.}"
 
-    if [ -e "${target}" ]; then
+    if [ -e "${target}" ] && [ ! -L "${target}" ]; then
         mv $target $target.bak
     fi
